+++ src/tiles/components/Hex.h
@@ -19,17 +19,17 @@ class Hex
 {
 public:
   Hex();
-  Hex(const int8_t q, const int8_t r);
+  Hex(const int q, const int r);
   Hex(const Hex &other);
   virtual ~Hex();
 
-  inline int8_t q() { return m_p_q; }
-  inline int8_t r() { return m_p_r; }
-  inline int8_t s() { return -m_p_q - m_p_r; }
+  inline int q() { return m_p_q; }
+  inline int r() { return m_p_r; }
+  inline int s() { return -m_p_q - m_p_r; }
 
-  inline int8_t q() const { return m_p_q; }
-  inline int8_t r() const { return m_p_r; }
-  inline int8_t s() const { return -m_p_q - m_p_r; }
+  inline int q() const { return m_p_q; }
+  inline int r() const { return m_p_r; }
+  inline int s() const { return -m_p_q - m_p_r; }
 
   Hex operator=(const Hex &other);
   Hex operator+(const Hex &other) const;
@@ -54,17 +54,17 @@ public:
   /// Returns coordinates for moving input amount along q axis
   /// @param[in] amount
   /// @return The resulting coordinates
-  inline Hex move_q(const int8_t amount) { return Hex(m_p_q, m_p_r - amount); }
+  inline Hex move_q(const int amount) { return Hex(m_p_q, m_p_r - amount); }
 
   /// Returns coordinates for moving input amount along r axis
   /// @param[in] amount
   /// @return The resulting coordinates
-  inline Hex move_r(const int8_t amount) { return Hex(m_p_q + amount, m_p_r); }
+  inline Hex move_r(const int amount) { return Hex(m_p_q + amount, m_p_r); }
 
   /// Returns coordinates for moving input amount along s axis
   /// @param[in] amount
   /// @return The resulting coordinates
-  inline Hex move_s(const int8_t amount)
+  inline Hex move_s(const int amount)
   {
     return Hex(m_p_q - amount, m_p_r + amount);
   }
@@ -75,8 +75,8 @@ public:
   /// Direction.
   inline Hex neighbor(const Direction d)
   {
-    int8_t q = m_p_q;
-    int8_t r = m_p_r;
+    int q = m_p_q;
+    int r = m_p_r;
     switch (d)
     {
     case Direction::north_west:
@@ -112,8 +112,8 @@ public:
   /// Direction.
   inline Hex neighbor(const Direction d) const
   {
-    int8_t q = m_p_q;
-    int8_t r = m_p_r;
+    int q = m_p_q;
+    int r = m_p_r;
     switch (d)
     {
     case Direction::north_west:
@@ -154,8 +154,8 @@ public:
   friend std::ostream &operator<<(std::ostream &os, const Hex &pos);
 
 protected:
-  int8_t m_p_q;
-  int8_t m_p_r;
+  int m_p_q;
+  int m_p_r;
 
 private:
 };
@@ -174,8 +174,8 @@ template <> struct std::hash<tile::Hex>
 {
   std::size_t operator()(tile::Hex const &hp) const noexcept
   {
-    std::size_t hq = std::hash<int8_t>{}(hp.q());
-    std::size_t hr = std::hash<int8_t>{}(hp.r());
+    std::size_t hq = std::hash<int>{}(hp.q());
+    std::size_t hr = std::hash<int>{}(hp.r());
     return hq ^
            (hr << 1); // or use boost::hash_combine (see Discussion)
                       // https://en.cppreference.com/w/Talk:cpp/utility/hash
